Ajuste de nombres de metodos a snake_case
